Option  --install-prefix (#60)
* Allow install of mirgecom to user-specified location

* Add specification of requirements file path

* Add required arguments in worflow checks

* Correct typo in yaml

* Try to fix makezip

* Update fetch-mirgecom.sh


* Update fetch-mirgecom.sh


* Update fetch-mirgecom.sh


* Update fetch-mirgecom.sh


* Update install-src-package.sh


* Update install-src-package.sh


* Update install.sh


* Update install.sh


* Update install.sh


* Add usage info to fetch-mirgecom

* Add usage info

* Update from PR suggested changes by MD

* Fix typo in var set

* Add missing argument

* Debug makezip tests

* Debug makezip tests

* Debug makezip tests

* Debug makezip tests

* Test more robust argument handling in install scripts"

* Fix up scripts - removed debugging, stale code

* Update install-pip-dependencies.sh


* remove shellcheck warnings

* remove commented code

* fix install

* fix

+++ makezip.sh
@@ -2,10 +2,20 @@
 
 set -o errexit -o nounset
 
-source ./parse_requirements.sh
-parse_requirements
+origin=$(pwd)
+requirements_file="${1-$origin}"
+install_loc="${2-$origin}"
 
-zipfile=$PWD/modules.zip
+if [ ! -f "$requirements_file" ]
+then
+    echo "makezip.sh::Error: Requirements file ($requirements_file) does not exist."
+    exit 1
+fi
+
+source ./parse_requirements.sh
+parse_requirements "$requirements_file"
+
+zipfile=$install_loc/modules.zip
 
 rm -f "$zipfile"
 
@@ -19,14 +29,14 @@ for i in "${!module_names[@]}"; do
     [[ -z $url ]] && continue
 
     # Skip non-Python submodules
-    [[ -f "$name/setup.py" ]] || continue
+    [[ -f "$install_loc/$name/setup.py" ]] || continue
 
     MY_MODULES+="$name "
 
-    cd "$name" || exit 2
+    cd "$install_loc/$name"
     echo "=== Zipping $name"
     zip -r "$zipfile" "$name/"
-    cd ..
+    cd "$origin"
 done
 
 MY_PYTHON=$(command -v python3)
@@ -35,7 +45,10 @@ echo "=== Preparing path file of '$MY_PYTHON'"
 echo "=== for importing modules from '$zipfile'"
 echo
 
-sitefile="$($MY_PYTHON -c 'import site; print(site.getsitepackages()[0])')/emirge.pth"
+sitepath="$($MY_PYTHON -c 'import site; print(site.getsitepackages()[0])')"
+sitefile="$sitepath/emirge.pth"
+echo "Found site path: $sitepath"
+echo "Attempting to make site file: $sitefile"
 
 echo "$zipfile" > "$sitefile"
